requirements/requirements-dev.txt
# Development dependencies for AutoJobAgent

# Core dependencies (from requirements.txt)
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
pytest-xdist>=3.3.1
pytest-html>=3.2.0
pytest-benchmark>=4.0.0

# Code quality
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
pylint>=2.17.0
bandit>=1.7.5
safety>=2.3.0

# Pre-commit hooks
pre-commit>=3.3.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0
sphinx-autodoc-typehints>=1.24.0

# Type stubs
types-requests>=2.31.0
types-beautifulsoup4>=4.12.0
types-openpyxl>=3.1.0
types-Pillow>=10.0.0

# Development tools
ipython>=8.14.0
jupyter>=1.0.0
notebook>=7.0.0

# Debugging
pdb++>=0.10.3
ipdb>=0.13.13
pudb>=2022.1.3

# Performance profiling
memory-profiler>=0.61.0
line-profiler>=4.1.0
py-spy>=0.3.14

# API testing
httpx>=0.24.0
respx>=0.20.0

# Browser testing
playwright-pytest>=0.4.0

# Mock services
responses>=0.23.0
requests-mock>=1.11.0

# Database testing (if needed)
pytest-postgresql>=5.0.0
pytest-redis>=3.0.0

# Load testing
locust>=2.16.0

# Security testing
semgrep>=1.35.0

# Dependency management
pip-tools>=7.1.0
pipdeptree>=2.13.0

# Build tools
build>=0.10.0
twine>=4.0.0
wheel>=0.41.0

# Environment management
python-dotenv>=1.0.0

# Git hooks
gitpython>=3.1.0

# Monitoring
psutil>=5.9.0

# Data validation
pydantic>=2.1.0
marshmallow>=3.20.0

# Configuration
pyyaml>=6.0.1
toml>=0.10.2

# Utilities
click>=8.1.0
rich>=13.5.0
tqdm>=4.66.0

# Development server
watchdog>=3.0.0
livereload>=2.6.3
